[libc] rename libc.c as stdio_init.c
