+++ components/libc/compilers/newlib/syscalls.c
@@ -225,7 +225,7 @@ _ssize_t _read_r(struct _reent *ptr, int fd, void *buf, size_t nbytes)
     if (fd == STDIN_FILENO)
     {
 #ifdef RT_USING_POSIX_STDIO
-        if (libc_stdio_get_console() < 0)
+        if (posix_stdio_get_console() < 0)
         {
             LOG_W("Do not invoke standard input before initializing Compiler");
             return 0;
