+++ components/libc/posix/io/stdio/SConscript
@@ -1,22 +1,9 @@
-# RT-Thread building script for component
-
-import os
 from building import *
 
-src     = []
+src     = ['stdio_init.c']
 cwd     = GetCurrentDir()
 CPPPATH = [cwd]
-group   = []
 
-if GetDepend('RT_USING_POSIX_STDIO'):
-    src += ['libc.c']
-
-group = DefineGroup('POSIX', src, depend = [''], CPPPATH = CPPPATH)
-
-list = os.listdir(cwd)
-for d in list:
-    path = os.path.join(cwd, d)
-    if os.path.isfile(os.path.join(path, 'SConscript')):
-        group = group + SConscript(os.path.join(d, 'SConscript'))
+group = DefineGroup('POSIX', src, depend = ['RT_USING_POSIX_STDIO'], CPPPATH = CPPPATH)
 
 Return('group')
