+++ components/libc/posix/io/stdio/stdio_init.h
@@ -6,10 +6,11 @@
  * Change Logs:
  * Date           Author       Notes
  * 2017/10/15     bernard      the first version
+ * 2023/08/06     Meco Man     rename as stdio_init.h
  */
 
-#ifndef __RTT_LIBC_H__
-#define __RTT_LIBC_H__
+#ifndef __POSIX_STDIO_INIT_H__
+#define __POSIX_STDIO_INIT_H__
 
 #include <rtconfig.h>
 
@@ -17,14 +18,14 @@
 extern "C" {
 #endif
 
-int libc_system_init(void);
+int posix_stdio_init(void);
 #ifdef RT_USING_POSIX_STDIO
-int libc_stdio_get_console(void);
-int libc_stdio_set_console(const char* device_name, int mode);
+int posix_stdio_get_console(void);
+int posix_stdio_set_console(const char* device_name, int mode);
 #endif /* RT_USING_POSIX_STDIO */
 
 #ifdef __cplusplus
 }
 #endif
 
-#endif
+#endif /* __POSIX_STDIO_INIT_H__ */
